Final Build
I am not proud of this, it has many bugs and I am yet to have gotten to the AI portion of the game which I intended to do. Right now there is a buggy pause menu and a Demo Menu with so much jank I decided to go with the original menu so the game is somewhat playable still.